refactor(posts): refactorizar y mover componentes PostDetailProfileImage, PostDetailHeader, PostDetailDescription, PostDetailMedia de archivo post-detail a carpeta features/posts/components
